Modified the variance calculation and least squares soln
The previous code appeared to use the standard deviation of the response as
the estimate of the variance.  This commit instead uses the maximum
likelihood estimate of the variance.

This commit also modifies the approach used to solve the least squares problem.
Rather than forming the normal matrix, the backslash operator is directly used.